apps/runners/pyproject.toml
[project]
name = "evalap-runners"
version = "0.1.0"
description = "EvalAP Background Runners"
requires-python = ">=3.12"
dependencies = [
    "evalap-core",
    "pyzmq~=26.2.0",
    "supervisor~=4.2.5",
    "pandas~=2.2.3",
]

[tool.uv.sources]
evalap-core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/evalap"]
